changed from float to int

+++ Projects/Project_3/Code/main.py
@@ -21,13 +21,13 @@ np.random.seed(12)
 # system size
 
 #num_layers = 1 
-num_layers = np.float(sys.argv[1])
+num_layers = np.int(sys.argv[1])
 #num_nodes = 300 
-num_nodes =  np.float(sys.argv[2])
+num_nodes =  np.int(sys.argv[2])
 #batch_size = 100 
-batch_size = np.float(sys.argv[3])
+batch_size = np.int(sys.argv[3])
 #epochs = 100 
-epochs = np.float(sys.argv[4])
+epochs = np.int(sys.argv[4])
 
 data = sys.argv[5]
 #data = input('Which dataset do you want to run, HIGGS or SUSY? [h/s]')
